Time signleton
Fixed time singleton and fixed on-screen  debug messages

+++ Source/Doggy_Vacation/Steak_Item.h
@@ -23,7 +23,7 @@ public:
 protected:
 
 	UPROPERTY(EditAnywhere) 
-	int Time_Change;
+	float Time_Change;
 	
 	UPROPERTY() 
 	int Heal;
